More optimizations in bandwidth.

+++ MinServerTest/src/server/Address.java
@@ -6,10 +6,10 @@ public class Address {
 	
 	public InetAddress address;
 	public int port;
-	public int check;
+	public byte check;
 	public byte[] fullData;
 	
-	public Address(InetAddress address, int port, int check, byte[] fullData) {
+	public Address(InetAddress address, int port, byte check, byte[] fullData) {
 		this.address = address;
 		this.port = port;
 		this.check = check;
