+++ MinServerTest/src/shared/ServerMessage.java
@@ -2,9 +2,9 @@ package shared;
 
 public class ServerMessage {
 	
-	public static final int NORMALOP = 0;
-	public static final int OUTOFSYNC = 1;
-	public static final int STATEREQUEST = 2;
-	public static final int IDASSIGN = 3;
+	public static final byte NORMALOP = 0;
+	public static final byte OUTOFSYNC = 1;
+	public static final byte STATEREQUEST = 2;
+	public static final byte IDASSIGN = 3;
 
 }
